% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSpower2P.R
\name{SSpower2P}
\alias{SSpower2P}
\title{Self-Starting Nls 'power2P' Regression Model}
\usage{
SSpower2P(predictor, a, b)
}
\arguments{
\item{predictor}{a numeric vector of values at which to evaluate the model.}

\item{a, b}{The numeric parameters responding to the exp2P model.}
}
\description{
This selfStart model evaluates the power regression function (formula as: y=a*x^b). It has an initial attribute that will evaluate initial estimates of the parameters 'a' and 'b' for a given set of data.
}
\examples{
library(ggtrendline)
x<-1:5
y<-c(2,4,8,20,25)
xy<-data.frame(x,y)
getInitial(y ~ SSpower2P(x,a,b), data = xy)
## Initial values are in fact the converged values

fitpower2P <- nls(y~SSpower2P(x,a,b), data=xy)
summary(fitpower2P)

prediction <- predFit(fitpower2P , data.frame(x=x), se.fit = TRUE,
                          level = 0.95, interval = "confidence")
yfitpower2P <- prediction$fit
yfitpower2P # output a matrix of predictions and bounds with column names fit, lwr, and upr.

}
\seealso{
\code{\link{ggtrendline}}, \code{\link{SSexp3P}}, \code{\link{SSpower3P}}, \code{\link[stats]{nls}}, \code{\link[stats]{selfStart}}
}
